Add files via upload
Fixed all returns on functions to only return the sorted data and moved the time start and stops to the line before and after the sort functions are run. Fixed a bug inside of average case complexity for the list generation. Initially it was generating an empty list. Tested all of the functions with all of the case complexities, and everything appeared to be working properly! :)