Add files via upload
Formatted a few spots throughout the program to make the I/O more directly match what was given to us in the sample in the project rubric. Notably added an f string where it displays the case complexity options - it now displays 4. Exit "insert sort func name" test and changes for each function that is selected.